media: move picture function from dive.c to picture.c
Currently, move only those functions that do not access dive
structures.
