+++ backend-shared/exportfuncs.cpp
@@ -10,6 +10,7 @@
 #include "core/errorhelper.h"
 #include "core/divefilter.h"
 #include "core/divesite.h"
+#include "core/picture.h"
 #include "exportfuncs.h"
 
 #if !defined(SUBSURFACE_MOBILE)
