+++ core/save-xml.c
@@ -20,6 +20,7 @@
 #include "device.h"
 #include "file.h"
 #include "membuffer.h"
+#include "picture.h"
 #include "strndup.h"
 #include "git-access.h"
 #include "qthelper.h"
